+++ test/e2e/stock-time-series/stock-time-series.test.ts
@@ -46,4 +46,22 @@ describe('StockTimeSeries', () => {
       expect(hasError).toBe(false);
     });
   });
+
+  describe('#dailyAdjusted', () => {
+    it('should make a request to search endpoint', async () => {
+      let stockTimeSeries = new StockTimeSeries(api);
+
+      let hasError = false;
+      try {
+        await stockTimeSeries.dailyAdjusted({
+          symbol: 'IBM',
+        });
+      } catch (err) {
+        console.error(err);
+        hasError = true;
+      }
+
+      expect(hasError).toBe(false);
+    });
+  });
 });