+++ test/unit/stock-time-series/stock-time-series.test.ts
@@ -14,6 +14,7 @@ describe('StockTimeSeries', () => {
   let api: AxiosInstance;
   let intradayData = {};
   let searchData = {};
+  let dailyAdjustedData = {};
 
   beforeEach(() => {
     api = axios.create({
@@ -58,6 +59,28 @@ describe('StockTimeSeries', () => {
           '9. matchScore': '0.7273',
         },
       ],
+    };
+
+    dailyAdjustedData = {
+      'Meta Data': {
+        '1. Information': 'Daily Time Series with Splits and Dividend Events',
+        '2. Symbol': 'IBM',
+        '3. Last Refreshed': '2021-10-15',
+        '4. Output Size': 'Compact',
+        '5. Time Zone': 'US/Eastern',
+      },
+      'Time Series (Daily)': {
+        '2021-10-15': {
+          '1. open': '143.39',
+          '2. high': '144.85',
+          '3. low': '142.79',
+          '4. close': '144.61',
+          '5. adjusted close': '144.61',
+          '6. volume': '3170857',
+          '7. dividend amount': '0.0000',
+          '8. split coefficient': '1.0',
+        },
+      },
     };
   });
 
@@ -236,6 +259,97 @@ describe('StockTimeSeries', () => {
       };
 
       const result = await stockTimeSeries.search(searchDTO);
+
+      expect(result).toEqual(csvData);
+    });
+  });
+
+  describe('#dailyAdjusted', () => {
+    it('should make a request to daily adjusted endpoint', async () => {
+      api.get = jest.fn().mockResolvedValue({ data: dailyAdjustedData });
+
+      const dailyAdjustedDTO = {
+        symbol: 'IBM',
+        outputsize: OutputSize.COMPACT,
+        datatype: DataType.JSON,
+      };
+
+      await stockTimeSeries.dailyAdjusted(dailyAdjustedDTO);
+
+      expect(api.get).toBeCalledWith('/query', {
+        params: {
+          ...dailyAdjustedDTO,
+          function: Function.TIME_SERIES_DAILY_ADJUSTED,
+        },
+      });
+    });
+
+    it('should return parsed daily adjusted data', async () => {
+      api.get = jest.fn().mockResolvedValue({ data: dailyAdjustedData });
+
+      const dailyAdjustedDTO = {
+        symbol: 'IBM',
+        outputsize: OutputSize.COMPACT,
+        datatype: DataType.JSON,
+      };
+
+      const result = await stockTimeSeries.dailyAdjusted(dailyAdjustedDTO);
+
+      expect(result.metadata.symbol).toEqual('IBM');
+      expect(result.metadata.outputSize).toEqual('Compact');
+      expect(result.timeSeries['2021-10-15']).toEqual({
+        open: '143.39',
+        high: '144.85',
+        low: '142.79',
+        close: '144.61',
+        adjustedClose: '144.61',
+        volume: '3170857',
+        dividendAmount: '0.0000',
+        splitCoefficient: '1.0',
+      });
+    });
+
+    it('should fail to parse daily adjusted data', async () => {
+      api.get = jest.fn().mockResolvedValue({});
+
+      try {
+        await stockTimeSeries.dailyAdjusted({
+          symbol: 'IBM',
+        });
+        fail('should have thrown an error');
+      } catch (err) {
+        expect(err).toBeInstanceOf(ParseResponseError);
+      }
+    });
+
+    it('should throw AlphaVantageRequestError because it fail to request daily adjusted data', async () => {
+      api.get = jest.fn().mockRejectedValue(new Error('some error'));
+
+      try {
+        await stockTimeSeries.dailyAdjusted({
+          symbol: 'IBM',
+        });
+        fail('should have thrown an error');
+      } catch (err) {
+        expect(err).toBeInstanceOf(AlphaVantageRequestError);
+      }
+    });
+
+    it('should return csv data', async () => {
+      const csvData =
+        'timestamp,open,high,low,close,adjusted_close,volume,dividend_amount,split_coefficient\
+        2021-10-15,143.39,144.85,142.79,144.61,144.61,3170857,0.0000,1.0\
+        2021-10-14,141.04,143.92,141.01,143.39,143.39,4217305,0.0000,1.0';
+
+      api.get = jest.fn().mockResolvedValue({ data: csvData });
+
+      const dailyAdjustedDTO = {
+        symbol: 'IBM',
+        outputsize: OutputSize.COMPACT,
+        datatype: DataType.CSV,
+      };
+
+      const result = await stockTimeSeries.dailyAdjusted(dailyAdjustedDTO);
 
       expect(result).toEqual(csvData);
     });